Add Home Model
- Add Home model data, mapping to domain model
- Optimize forecast domain model, fix wrong serializedname

+++ MVIWeather/weather-stack/remote-repo-weather/src/main/java/com/example/mvi/weather/remoterepo/weather/model/forecast/City.kt
@@ -21,7 +21,7 @@ data class City(
     val population: Int,
 
     @SerializedName("timezone")
-    val timezone: Int,
+    val timezone: Long,
 
     @SerializedName("sunrise")
     val sunrise: Long,
